% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.dnorm.r
\name{mstep.dnorm}
\alias{mstep.dnorm}
\title{M-step function for univariate normal distribution}
\usage{
mstep.dnorm(obs, cond, ctrl, mean, sd)
}
\arguments{
\item{obs}{numerical vector of observations}

\item{cond}{estimates of the conditional expectations as computed by \link{thmm.estep}}

\item{ctrl}{list of control parameters created by \link{thmm.bw.ctrl}}

\item{mean}{list of states means (may be missing)}

\item{sd}{list of states sd (may be missing)}
}
\value{
list of updated parameters mean and sd
}
\description{
used as the M-step function of \link{thmm.baumwelch} for
univariate normal distribution.
}
\seealso{
\link{thmm.baumwelch}
}
