% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.init.r
\name{print.DTHmm}
\alias{print.DTHmm}
\title{print method for DTHmm}
\usage{
\method{print}{DTHmm}(x, ...)
}
\arguments{
\item{x}{a DTHmm object}

\item{...}{further arguments passed to or from other methods}
}
\value{
invisible(x)
}
\description{
print method for DTHmm
}
