% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.bic.r
\name{thmm.bic}
\alias{thmm.bic}
\title{compute BIC}
\usage{
thmm.bic(hmm, obs, np, nt = nrow(hmm$trans) * (nrow(hmm$trans) - 1),
  ni = nrow(hmm$trans) - 1, llk = NA, .useC = TRUE)
}
\arguments{
\item{hmm}{DTHmm model (from \link{thmm.init})}

\item{obs}{numerical vector of observations}

\item{np}{number of free parameters of hmm density function (e.g. 2 for normal, 1 for poisson).}

\item{nt}{number of free parameters for transition matrix.}

\item{ni}{number of free parameters for initial probabilities.}

\item{llk}{numeric, log-likelihood(obs | hmm). if NA then llk will be computed using \link{thmm.forward}.}

\item{.useC}{logical if TRUE (default) use C code else use R code for \link{thmm.forward}}
}
\value{
numeric BIC
}
\description{
compute Bayesian Information Criterion (BIC) of hmm
and observation.
}
\details{
BIC = \code{-2 * llk + (nt + ni + M * np) * log(N)}\cr with
\code{M=number of states; N=length(obs)}
}
\examples{
hmm <- thmm.init(dnorm, 0.1, mean=c(-1,1), sd=0.1)
obs <- thmm.simulate(hmm, 100, .seed=0)$values
thmm.bic(hmm, obs, np=2, nt=1, ni=0)

}
