% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.baumwelch.r
\name{thmm.bw.ctrl}
\alias{thmm.bw.ctrl}
\title{control parameters for Baum-Welch algorithm}
\usage{
thmm.bw.ctrl(hmm, maxiter = NA, tol = 1e-05, verbose = FALSE,
  converge = expression(difLL <= ctrl$tol), do.trans = TRUE,
  do.init = TRUE, constraint = NULL, ...)
}
\arguments{
\item{hmm}{DTHmm model (from \link{thmm.init})}

\item{maxiter}{integer max number of iteration (NA will set to hmm
dependent value)}

\item{tol}{numeric convergence criterion (delta_log-likelihood)}

\item{verbose}{logical be verbose}

\item{converge}{expression of the convergence criterion}

\item{do.trans}{logical do optimize transition matrix}

\item{do.init}{logical do optimize initial probabilities}

\item{constraint}{if not NULL, should contain a function receiving a list of
optimized parameters (at each M-step) and should return a list (with
the same format) of constrained parameters values (see examples).}

\item{...}{any parameter passed to \code{<density>.mstep} function
(see \link{thmm.baumwelch}). in particular:
\itemize{
\item parameters of the form
\code{do.<parameter>} are interpreted as logical controling whether
or not we should optimize the <parameter>.
note that to prevent update you should explicitely set \code{do.<parameter>}
to \code{FALSE}. if \code{do.<parameter>} is absent (or set to TRUE) then
\code{<parameter>} \bold{will be} updated.
\item parameters of the form \code{upper|lower.<parameter>}
are interpreted as numerical upper/lower bounds. if absent then
no bounds (i.e. [-Inf, +Inf]) are assumed.
}}
}
\value{
named list of control parameters for \link{thmm.baumwelch}
}
\description{
make a list of parameters to control
\link{thmm.baumwelch} execution.
}
\note{
\code{constraint} provides a more general form of control than
\code{do.<parameter>, upper|lower.<parameter>} but may be less efficient
since it is called right after the mstep function.

(for developpers) \code{do.<parameter>, upper|lower.<parameter>}
should be implemented for every mstep function. see \link{mstep.dnorm} as an exemple.
\code{constraint} is called by \link{thmm.baumwelch} directly.
}
\seealso{
\link{thmm.baumwelch}
}
