% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.baumwelch.r
\name{thmm.estep}
\alias{thmm.estep}
\title{E-Step of EM Baum-Welch algorithm}
\usage{
thmm.estep(hmm, obs, .useC = TRUE)
}
\arguments{
\item{hmm}{DTHmm model (from \link{thmm.init})}

\item{obs}{numerical vector of observations}

\item{.useC}{logical if TRUE (default) use C code else use R code}
}
\value{
list with three components:\cr
\itemize{
\item u length(obs)*number_states matrix containing estimates of the
conditional expectations (see details)
\item v length(obs)*number_states*number_states matrix containing estimates of the
conditional expectations (see details)
\item loglike log-likelihood
}
}
\description{
performs the expectation step of the EM algorithm
for a discrete time HMM process.
this function is called by the \link{thmm.baumwelch}.
}
\details{
u is defined as:\cr
\code{u{i,j}=Pr(state_i=j | obs, hmm)}\cr
that is the same as \code{gamma} in \link{thmm.forward.backward}

v is defined as:\cr
\code{v({i,j,k}=Pr(state_i-1=j, state_i=k | obs, hmm)}\cr
also known as \code{chsi}
}
\seealso{
\link{thmm.baumwelch}
}
