% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.forback.r
\name{thmm.forward.backward}
\alias{thmm.forward.backward}
\title{(log) forward/backward probabilities}
\usage{
thmm.forward.backward(hmm, obs, .useC = TRUE)
}
\arguments{
\item{hmm}{DTHmm model (from \link{thmm.init})}

\item{obs}{numerical vector of observations}

\item{.useC}{logical if TRUE (default) use C code else use R code}
}
\value{
list with five components: \code{logalpha}, \code{logbeta},
\code{gamma}, \code{rho}, \code{loglike}
\itemize{
\item\code{logalpha} : the (log of) forward probabilities (densities) matrix:\cr
\code{alpha_{i,j} = Pr{ X_1 = obs_1, ..., X_i = obs_i, state_i = j | hmm}}
that is the probability (density) of seeing the partial sequence
\code{(obs_1, ..., obs_i)} and ending up in state j at time i for this
hmm.

\item\code{logbeta} : the (log of) backward probabilities (densities) matrix:\cr
\code{beta_{i,j} = Pr{ X_{i+1} = obs_{i+1}, ..., X_n = obs_n | state_i = j , hmm}}
that is the probability (density) of the ending partial sequence
\code{(obs_i+1, ..., obs_n)} given that we started at state j at time i,
for this hmm.

\item\code{gamma} : the probabilities matrix:\cr
\code{gamma_{i,j} = Pr{ state_i = j | obs , hmm}}
that is the probability of being at time i in state j given
this observation and this hmm.

\item\code{logrho} : the (log of) probabilities (densities) vector:\cr
\code{rho_{i} = Pr{ X_1 = obs_1, ..., X_i = obs_i | hmm}}
that is the probability (density) of seeing the partial sequence
\code{(obs_1, ..., obs_i)} for this hmm.

\item\code{loglike} : the log-likelihood of this observation with this hmm:\cr
\code{log(Pr{ X_1 = obs_1, ..., X_n = obs_n | hmm})} (see \link{thmm.forward})
}
}
\description{
compute forward/backward probabilities of observations
according to a discrete time hidden markov model (DTHmm).
}
\note{
the code takes care of rescaling values during calculation
to avoid underflow problems and is quite robust in practice.

by definition \code{loglike = logrho(length(obs))} (although it is not
computed this way).
}
\examples{
hmm <- thmm.init(dnorm, 0.1, mean=c(-1,1), sd=0.1)
obs <- c(rnorm(100, -1, 0.1), rnorm(100, 1, 0.1))
fb <- thmm.forward.backward(hmm, obs)
#
# test numerical stability
#
obs[101] <- 1000000  # this may creates overflow
fb <- thmm.forward.backward(hmm, obs)
\dontrun{
plot(fb$gamma[,2])}
}
\seealso{
\link{thmm.forward}, \link{thmm.backward}
}
