% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.init.r
\name{thmm.parameters}
\alias{thmm.parameters}
\title{get hmm parameters}
\usage{
thmm.parameters(hmm, what = NULL)
}
\arguments{
\item{hmm}{DTHmm model (from \link{thmm.init})}

\item{what}{character parameter name to retrieve (all parameters if NULL (default))}
}
\value{
if \code{what==NULL} return a list of 5 components: \code{nstates},
\code{density}, \code{args}, \code{init}, \code{trans}
\itemize{
\item nstates : integer number of states
\item density : character string name of probability density function
\item args : list (of length nstates) of named lists, each element
gives the density function parameters for state i. (see note).
\item init : numerical vector of initial probabilities
\item trans : numerical transition matrix
}
if \code{what!=NULL} then what should be either one of the above names or
the name of one of the density function parameters. the function then return
only this information.
}
\description{
get hmm global and individual states parameters
}
\examples{
# univariate gaussian hmm
hmm <- thmm.init(dnorm, 0.1, mean=c(-1,1), sd=1)
ns <- thmm.parameters(hmm, "nstates")
means <- thmm.parameters(hmm, "mean")
# viterbi decode
obs <- rnorm(100)
vit <- thmm.viterbi(hmm, obs)
# as values
means[vit$states]

}
